# docs/swagger.yaml
basePath: /
definitions:
  model.AudioMessage:
    properties:
      audio:
        type: string
      id:
        type: string
      mimetype:
        type: string
      phone:
        type: string
      ptt:
        type: boolean
    type: object
  model.ContactMessage:
    properties:
      contact_name:
        type: string
      contact_vcard:
        type: string
      id:
        type: string
      phone:
        type: string
    type: object
  model.DeleteMessage:
    properties:
      message_id:
        type: string
      phone:
        type: string
    type: object
  model.DocumentMessage:
    properties:
      caption:
        type: string
      document:
        type: string
      filename:
        type: string
      id:
        type: string
      mimetype:
        type: string
      phone:
        type: string
    type: object
  model.ImageMessage:
    properties:
      caption:
        type: string
      id:
        type: string
      image:
        type: string
      mimetype:
        type: string
      phone:
        type: string
    type: object
  model.LocationMessage:
    properties:
      address:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
    type: object
...
  model.ReactionMessage:
    properties:
      emoji:
        type: string
      message_id:
        type: string
      phone:
        type: string
    type: object
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  model.TextMessage:
    properties:
      id:
        type: string
      message:
        type: string
      phone:
        type: string
    type: object
...
  model.VideoMessage:
    properties:
      caption:
        type: string
      id:
        type: string
      mimetype:
        type: string
      phone:
        type: string
      video:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@fiozap.io
    name: FioZap Support
  description: WhatsApp API using whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FioZap API
  version: "1.0"
paths:
...
  /chat/delete:
    post:
      consumes:
      - application/json
      description: Delete a sent message
      parameters:
      - description: Delete data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.DeleteMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete message
      tags:
      - Messages
...
  /chat/react:
    post:
      consumes:
      - application/json
      description: Send a reaction to a message
      parameters:
      - description: Reaction data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.ReactionMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: React to message
      tags:
      - Messages
  /chat/send/audio:
    post:
      consumes:
      - application/json
      description: Send an audio file to a phone number
      parameters:
      - description: Audio data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.AudioMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send audio
      tags:
      - Messages
  /chat/send/contact:
    post:
      consumes:
      - application/json
      description: Send a contact card to a phone number
      parameters:
      - description: Contact data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.ContactMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send contact
      tags:
      - Messages
  /chat/send/document:
    post:
      consumes:
      - application/json
      description: Send a document to a phone number
      parameters:
      - description: Document data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.DocumentMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send document
      tags:
      - Messages
  /chat/send/image:
    post:
      consumes:
      - application/json
      description: Send an image to a phone number
      parameters:
      - description: Image data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.ImageMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send image
      tags:
      - Messages
  /chat/send/location:
    post:
      consumes:
      - application/json
      description: Send a location to a phone number
      parameters:
      - description: Location data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.LocationMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send location
      tags:
      - Messages
  /chat/send/text:
    post:
      consumes:
      - application/json
      description: Send a text message to a phone number
      parameters:
      - description: Message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.TextMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send text message
      tags:
      - Messages
  /chat/send/video:
    post:
      consumes:
      - application/json
      description: Send a video to a phone number
      parameters:
      - description: Video data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.VideoMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send video
      tags:
      - Messages
...
securityDefinitions:
  AdminKeyAuth:
    description: Admin token for user management
    in: header
    name: Authorization
    type: apiKey
  ApiKeyAuth:
    description: User API token for session operations
    in: header
    name: Token
    type: apiKey
swagger: "2.0"